% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyGREAT.R
\name{applyGREAT}
\alias{applyGREAT}
\title{Find Enrichment of GO Terms in PatternMarker Peaks using GREAT}
\usage{
applyGREAT(
  cogapsResult,
  granges,
  genome,
  scoreThreshold = NULL,
  GREATCategory = "GO"
)
}
\arguments{
\item{cogapsResult}{result object from CoGAPS}

\item{granges}{GRanges object corresponding to the peaks of the scATAC-seq
data CoGAPS was applied to}

\item{genome}{UCSC genome designation for input to the sumbitGreatJob
function from the rGREAT package (e.g. "hg19")}

\item{scoreThreshold}{threshold of PatternMarker score to take peaks for
analysis, higher values return more peaks. Defaults to use all PatternMarker
genes with value NULL}

\item{GREATCategory}{input to the category argument of the rGREAT 
getEnrichmentTables function. Usually "GO" or "Genes"}
}
\value{
list containing enrichment results for each pattern
}
\description{
Use the rGREAT package to find enrichment of GO terms or genes for the peaks found to
be most pattern differentiating using the PatternMarker statistic.
}
\examples{
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/PaperVersion/data/schepCogapsResult.rda?raw=true")
data(schepGranges)

GOenrichment <- applyGREAT(cogapsResult = schepCogapsResult, granges = schepGranges, genome = "hg19")
}
