% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACTransferLearning.R
\name{ATACTransferLearning}
\alias{ATACTransferLearning}
\title{Transfer Learning between ATACseq data sets using projectR}
\usage{
ATACTransferLearning(
  newData,
  CoGAPSResult,
  originalPeaks,
  originalGranges,
  newGranges
)
}
\arguments{
\item{newData}{the ATAC data to project into}

\item{CoGAPSResult}{result from CoGAPS run on original ATAC data}

\item{originalPeaks}{peaks from the ATAC data Cogaps was run on}

\item{originalGranges}{granges of the peaks for the data set Cogaps was run
on}

\item{newGranges}{granges of the peaks for the new data set}
}
\value{
A matrix of the projected patterns in the input data as well as
  p-values for each element of that matrix.
}
\description{
Wrapper function for projectR which finds overlaps between the peaks of the
atac data CoGAPS was run on and maps them to new data set the user wishes to
project learned patterns into.
}
\examples{
data(schepCogapsResult)
data(schepFilteredPeaks)
data(schepGranges)


#getting count matrix - peaks x cells
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/PaperVersion/BuenrostroFinalSubsetData.Rdata?raw=true")

#getting GRanges for peaks
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/PaperVersion/BuenrostroGRanges.Rdata?raw=true")

#getting celltypes
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/PaperVersion/BuenrostroCellTypes.Rdata?raw=true")

projectRResults <- ATACTransferLearning(newData = BuenrostroFinalSubsetData, CoGAPSResult = schepCogapsResult, originalPeaks = schepFilteredPeaks, originalGranges = schepGranges,newGranges = BuenrostroGRanges)
}
