% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgaps_plot.R
\name{cgapsPlot}
\alias{cgapsPlot}
\title{Plot Individual CoGAPS Patterns}
\usage{
cgapsPlot(
  cgaps_result,
  sample.classifier,
  cols = NULL,
  sort = TRUE,
  patterns = NULL,
  matrix = FALSE,
  seed = 42,
  ...
)
}
\arguments{
\item{cgaps_result}{CoGAPSResult object from a CoGAPS run or the pattern
matrix (matrix must be set equal to TRUE in the latter case)}

\item{sample.classifier}{factor of sample classifications for all cells for
the data to be plotted by (e.g. celltypes)}

\item{cols}{vector of colors to be used for the cell classes; should have
the same number of colors as levels of the sample.classifier factor. If
left null a list of colors is produced}

\item{sort}{TRUE if samples will be sorted according to sample.classifier
prior to plotting}

\item{patterns}{numerical vector of patterns to be plotted; if null all
patterns are plotted}

\item{matrix}{if false cgaps_result is interpreted as a CoGAPSResult object,
if true it is interpreted as the pattern matrix}

\item{seed}{random seed to generate colors; for reproducibility}

\item{...}{addition arguments to plot function}
}
\value{
Series of plots of pattern matrix patterns colored by cell
  classifications
}
\description{
Function to plot each pattern of the pattern matrix from a cogapsResult and
color by cell classifier information to identify which patterns identify
which cell classes.
}
\examples{
data("schepCogapsResult")
data(schepCellTypes)

cgapsPlot(schepCogapsResult, schepCellTypes)
}
