% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSubset_by_sparsity.R
\name{dataSubsetBySparsity}
\alias{dataSubsetBySparsity}
\title{Filter scATACseq by sparsity}
\usage{
dataSubsetBySparsity(data, cell_list, peak_list, cell_cut = 0.99,
  peak_cut = 0.99)
}
\arguments{
\item{data}{matrix of read counts peaks x cells}

\item{cell_list}{list of cell names/identifiers for the data}

\item{peak_list}{list of peaks from the data}

\item{cell_cut}{threshold of sparsity to filter at (eg. 0.99 filters all
cells with more than 99 percent zero values)}

\item{peak_cut}{threshold of sparsity to filter at for peaks}
}
\value{
nested list containing the subset data, a list of peaks, and list of
  cells
}
\description{
Function to filter a set of scATACseq data by sparsity and return a subset of
filtered data, as well as list of the remaining cells and peaks.
}
\examples{
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/master/schepATACPreFilter.Rdata?raw=true")
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/master/schepUnfilteredCells.Rdata?raw=true")
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/master/schepUnfilteredPeaks.Rdata?raw=true")

dataSubset = dataSubsetBySparsity(data = schepATACPreFilter, cell_list = schepUnfilteredCells, peak_list = schepUnfilteredPeaks, cell_cut = 0.99, peak_cut = 0.99)

}
