% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMotifList}
\alias{exampleMotifList}
\title{Example list of motifs for examples}
\format{
PWMMatrixList of length 100
}
\usage{
exampleMotifList
}
\description{
PWMMatrixList used for examples with functions based on DNA motifs.
Each entry contains the motif ID and the probability of each nucleotide
at each position, as a matrix.
}
\keyword{datasets}
