% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGOEnrichment.R
\name{findGOEnrichment}
\alias{findGOEnrichment}
\title{Find Enrichment of GO Terms in PatternMarker Peaks}
\usage{
findGOEnrichment(cogapsResult, granges, genome, scoreThreshold = NULL)
}
\arguments{
\item{cogapsResult}{result object from CoGAPS}

\item{granges}{GRanges object corresponding to the peaks of the scATAC-seq
data CoGAPS was applied to}

\item{genome}{UCSC genome designation for input to the sumbitGreatJob
function from the rGREAT package (e.g. "hg19")}

\item{scoreThreshold}{threshold of PatternMarker score to take peaks for
analysis, higher values return more peaks. Default is NULL, but rGREAT has
better performance on lower numbers of peaks/genes, so more stringent
thresholding will often improve results. Reasonable values generally range
from 0.05 to 0.25}
}
\value{
list containing GO enrichment result for each pattern
}
\description{
Use the rGREAT package to find enrichment of GO terms for the peaks found to
be most pattern differentiating using the PatternMarker statistic.
}
\examples{
data(schepCogapsResult)
data(schepGranges)

GOenrichment <- findGOEnrichment(cogapsResult = schepCogapsResult, granges = schepGranges, genome = "hg19", scoreThreshold = 0.05)
}
