% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldAccessibility.R
\name{foldAccessibility}
\alias{foldAccessibility}
\title{Estimate fold Accessibility of a Gene Relative to Average}
\usage{
foldAccessibility(peaksAccessibility, cellTypeList, cellType, binaryMatrix)
}
\arguments{
\item{peaksAccessibility}{the binarized accessibility of a set of peaks; one
value returned from the geneAccessibility function}

\item{cellTypeList}{list of celltypes grouping cells in the data}

\item{cellType}{the particular cell type of interest from within cellTypeList}

\item{binaryMatrix}{binarized scATAC data matrix}
}
\value{
Fold accessibility value as compared to average peaks for a given
  cell type
}
\description{
Compares the accessibility of peaks overlapping with a gene, as returned by
the geneAccessibility function to the average accessibility of peaks within a
given cell population. Meant to provide a rough estimate of how accessible a
gene is with values higher than 1 providing evidence of differential
accessibility (and thus implying possible transcription), with values lower
than 1 indicating the opposite.
}
\examples{
data("subsetSchepData")
data(schepCellTypes)
library(Homo.sapiens)
geneList <- c("TAL1", "IRF1")
data(schepGranges)
binarizedData <- (subsetSchepData > 0) + 0
accessiblePeaks <- geneAccessibility(geneList = geneList, peakGranges = schepGranges, atacData = subsetSchepData, genome = Homo.sapiens)
foldAccessibility(peaksAccessibility = accessiblePeaks$TAL1, cellTypeList = schepCellTypes, cellType = "K562 Erythroleukemia", binaryMatrix = binarizedData)
}
