% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAccessibility.R
\name{geneAccessibility}
\alias{geneAccessibility}
\title{Find the accessibility of the peaks overlapping a set of genes and their
promoters}
\usage{
geneAccessibility(geneList, peakGranges, atacData, genome)
}
\arguments{
\item{geneList}{vector of HGNC gene symbols to find overlapping peaks for in
the data}

\item{peakGranges}{a GRanges object corresponding to the peaks in the
atacData matrix, in the same order as the rows of the atacData matrix}

\item{atacData}{a single-cell ATAC-seq count matrix peaks by cells}

\item{genome}{TxDb object to produce gene GRanges from}
}
\value{
List of matrices corresponding to the accessible peaks overlapping
  with each gene across all cells in the data
}
\description{
The accessibility of a particular set of interest genes is checked by testing
overlap of peaks with the genes and gene promoters and then returning the
binarized accesibility data for those peaks
}
\examples{
library(Homo.sapiens)
geneList <- c("TAL1", "IRF1")
data(schepGranges)
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/PaperVersion/data/schepFilteredData.rda?raw=true")
accessiblePeaks <- geneAccessibility(geneList = geneList, peakGranges = schepGranges, atacData = schepFilteredData, genome = Homo.sapiens)
}
