% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_peak_matching.R
\name{genePatternMatch}
\alias{genePatternMatch}
\title{Match genes to pattern differentiating peaks}
\usage{
genePatternMatch(cogapsResult, numregions, generanges, genome)
}
\arguments{
\item{cogapsResult}{the CogapsResult object produced by a CoGAPS run}

\item{numregions}{the number of the most pattern-defining regions to
investigate for each pattern}

\item{generanges}{GRanges object corresponding to the genomic regions
identified as peaks for the ATAC-seq data that CoGAPS was run on}

\item{genome}{Homo.sapiens or Mus.musculus currently supported}
}
\value{
double nested list containing lists of the genes in, nearest, and
  following the top numregions for each pattern
}
\description{
Function to take as input CoGAPS results for ATAC-seq data and find genes
within the most "pattern-defining" regions (as identified by cut thresholded
pattern Marker statistic from the CoGAPS package), as well as the nearest
gene and the nearest gene following the region
}
\examples{
data(schepCogapsResult)
data(schepGranges)

genes = genePatternMatch(schepCogapsResult, 500, schepGranges, Homo.sapiens)
}
