% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGeneAccessibility.R
\name{heatmapGeneAccessibility}
\alias{heatmapGeneAccessibility}
\title{Heatmap Gene Accessibility}
\usage{
heatmapGeneAccessibility(
  genePeaks,
  celltypes,
  colColors = NULL,
  order = TRUE,
  seed = 42,
  ...
)
}
\arguments{
\item{genePeaks}{The peaks corresponding to a singular gene; one element of
the list ouptut by geneAccessibility()}

\item{celltypes}{List or factor of celltypes corresponding to the cells in
the scATAC-seq data set the peaks were found in}

\item{colColors}{A vector of colors to color the celltypes by, if NULL a
random vector of colors is generated}

\item{order}{should the data be ordered by the celltype classifier? TRUE by
default}

\item{seed}{random seed to generate colors if colColors is NULL}

\item{...}{additional arguments to the heatmap.2 function from the gplots
package}
}
\value{
A plot of the peaks overlapping with a particular gene of interest
}
\description{
Use the output from geneAccessibility function to plot a heatmap of the
accessible peaks for a particular gene.
}
\examples{
library(Homo.sapiens)
geneList <- c("TAL1", "EGR1")
data(schepGranges)
repmis::source_data("https://github.com/FertigLab/ATACCoGAPS/blob/PaperVersion/data/schepFilteredData.rda?raw=true")
data(schepCelltypes)
accessiblePeaks <- geneAccessibility(geneList = geneList, peakGranges = schepGranges, atacData = schepFilteredData, genome = Homo.sapiens)
heatmapGeneAccessibility(genePeaks = accessiblePeaks$TAL1, celltypes = schepCelltypes)
}
