% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_pattern_marker_regions.R
\name{heatmapPatternMarkers}
\alias{heatmapPatternMarkers}
\title{Create Heatmap of PatternMarker Peaks}
\usage{
heatmapPatternMarkers(cgaps_result, atac_data, celltypes, numregions,
  colColors = NULL, rowColors = NULL, patterns = NULL, seed = 42,
  order = TRUE, ...)
}
\arguments{
\item{cgaps_result}{CogapsResult object from CoGAPS run}

\item{atac_data}{a numeric matrix of the ATAC data input to CoGAPS}

\item{celltypes}{a list or factor of celltypes corresponding to the positions
of those cells in the atac_data matrix}

\item{numregions}{number of chromosomal regions to plot for each CoGAPS
pattern}

\item{colColors}{column-wise colors for distinguishing celltypes. If NULL,
will be generated randomly}

\item{rowColors}{row-wise colors for distinguishing patterns. If NULL will be
generated randomly}

\item{patterns}{which patterns should be plotted, if NULL all will be plotted}

\item{seed}{random seed for generating colors for plot; for reproducibility}

\item{order}{option whether to sort the data by celltype before plotting,
TRUE by default}

\item{...}{additional arguments to the heatmap.2 function}
}
\value{
heatmap of the accessibility for numregions for each pattern
}
\description{
Function to make a heatmap of the accessibility of the most differentially
accessible regions as discovered by CoGAPS.
}
\section{Note}{
 If you get the error:  "Error in plot.new() : figure margins
  too large" while using this function in RStudio just make the plotting pane
  in Rstudio larger and run the code again; this error only means the legend
  is being cut off in any case, the main plot will still appear correctly
}

\examples{
data(schepCogapsResult)
data(schepCelltypes)
data(schepFilteredData)

heatmapPatternMarkers(schepCogapsResult, atac_data = schepFilteredData,
                      celltypes = schepCelltypes, numregions = 50)
}
