% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapPatternMatrix.R
\name{heatmapPatternMatrix}
\alias{heatmapPatternMatrix}
\title{Plot the patternMatrix as a heatmap}
\usage{
heatmapPatternMatrix(
  cgaps_result,
  sample.classifier,
  cellCols = NULL,
  sort = TRUE,
  patterns = NULL,
  matrix = FALSE,
  rowColors = NULL,
  seed = 42,
  ...
)
}
\arguments{
\item{cgaps_result}{CoGAPSResult object from a CoGAPS run or the pattern
matrix (matrix must be set equal to TRUE in the latter case)}

\item{sample.classifier}{factor of sample classifications for all cells for
the data to be plotted by (e.g. celltypes)}

\item{cellCols}{vector of colors to be used for the cell classes; should have
the same number of colors as levels of the sample.classifier factor. If
left null a list of colors is produced}

\item{sort}{TRUE if samples will be sorted according to sample.classifier
prior to plotting}

\item{patterns}{numerical vector of patterns to be plotted; if null all
patterns are plotted}

\item{matrix}{if false cgaps_result is interpreted as a CoGAPSResult object,
if true it is interpreted as the pattern matrix being input directly}

\item{rowColors}{vector of colors to plot along patterns, if NULL generated automatically}

\item{seed}{random seed to generate colors; for reproducibility}

\item{...}{additional arguments to the heatmap.2 function}
}
\value{
Heatmap of patternMatrix with color labels for samples
}
\description{
Selects the pattermMatrix (patterns by cells) from the CoGAPSResult and plots
the data as a heatmap. Intended to visualize the celltypes distinguished by
the patterns found by CoGAPS.
}
\examples{
data("schepCogapsResult")
data(schepCellTypes)

heatmapPatternMatrix(schepCogapsResult, sample.classifier = schepCellTypes)
}
