% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifSummarization.R
\name{motifSummarization}
\alias{motifSummarization}
\title{Map Peaks to DNA motifs in scATAC-seq Data}
\usage{
motifSummarization(
  motifList,
  scATACData,
  granges,
  genome,
  cellNames,
  pCutoff = 5e-09
)
}
\arguments{
\item{motifList}{PWMatrixList object of motifs (from the TFBS tools package)}

\item{scATACData}{matrix of scATACseq data, peaks (rows) by cells (columns)}

\item{granges}{GenomicRanges object corresponding to all peaks used to summarize scATACData}

\item{genome}{The UCSC Genome to use for input to motifmatchr (e.g "hg19")}

\item{cellNames}{List of cellnames corresponding to the cells in scATACData}

\item{pCutoff}{p-value cutoff for motifmatchr, 5e-09 by default to identify only matches with high confidence}
}
\value{
matrix for input to CoGAPS with summary to motifs; motifs by cells
}
\description{
Provides functionality to summarize scATAC-seq data by motifs from peak
summary. Uses motifmatchr to prepare data for CoGAPS run using motif
summarization
}
\examples{
\dontrun{
motifSummTest = motifSummarization(motifList = motifs, scATACData = scatac, granges = peakGranges, genome = "hg19", cellNames = cells, pCutoff = 5e-09)
}
}
