% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePathwayMatchingV2.R
\name{pathwayMatch}
\alias{pathwayMatch}
\title{Matches list of genes to pathways}
\usage{
pathwayMatch(gene_list, pathways, p_threshold = 0.05,
  pAdjustMethod = "BH")
}
\arguments{
\item{gene_list}{Result from the genePatternMatch function, a list of genes
for each pattern}

\item{pathways}{List of pathways to perform gene enrichment on. Recommended
to download using msigdbr (see examples)}

\item{p_threshold}{significance level to use in enrichment analysis}

\item{pAdjustMethod}{multiple testing correction method to apply using the
p.adjust options (e.g. "BH")}
}
\value{
List of gene overlap objects, pathways with significant overlap and
  pathway names for each pattern
}
\description{
Takes the result of the genePatternMatch function and finds significantly
enriched pathways for each pattern.
}
\examples{
data(schepCogapsResult)
data(schepGranges)

genes = genePatternMatch(schepCogapsResult, 500, schepGranges, Homo.sapiens)

library(dplyr)
pathways = msigdbr::msigdbr(species = "Homo sapiens", category =
                             "H") \%>\% dplyr::select(gs_name, gene_symbol) \%>\% as.data.frame()

matchedPathways = pathwayMatch(genes, pathways, p_threshold = 0.001)
}
