% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternMarkerCellClassifier.R
\name{patternMarkerCellClassifier}
\alias{patternMarkerCellClassifier}
\title{Match cells to patterns}
\usage{
patternMarkerCellClassifier(cgapsResult)
}
\arguments{
\item{cgapsResult}{a CoGAPSResult object}
}
\value{
list containing a prediction matrix and vector classifying cells to patterns
}
\description{
Use the patternMarker statistic to determine which cells belong to each
pattern in the data
}
\examples{
data("schepCogapsResult")
pClass <- patternMarkerClassifier(schepCogapsResult)
}
