% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaksToGRanges.R
\name{peaksToGRanges}
\alias{peaksToGRanges}
\title{List of peaks to GRanges}
\usage{
peaksToGRanges(region_list, sep)
}
\arguments{
\item{region_list}{character list or vector of chromosomal regions/peaks in
form chromosomenumber(sep)start(sep)end eg. Chr1-345678-398744}

\item{sep}{separator between information pieces of string (conventionally "-"
or ".")}
}
\value{
GRanges corrsponding to input list of region information
}
\description{
Wrapper function for makeGrangesFromDataFrame() from the GenomicRanges
package to build GRanges objects from character list of chromosomal regions
because this is a common format to receive peak information.
}
\section{Note}{
 If region_list is a dataframe you should use the GenomicRanges
  function makeGRangesFromDataFrame which this function applies
}

\examples{
data(schepPeaks)

schepGranges = peaksToGRanges(schepPeaks, sep = "-")
}
