% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqTFValidation.R
\name{RNAseqTFValidation}
\alias{RNAseqTFValidation}
\title{Validate TF Findings with RNA-seq CoGAPS}
\usage{
RNAseqTFValidation(TFGenes, RNACoGAPSResult, ATACPatternSet, RNAPatternSet,
  matrix = FALSE)
}
\arguments{
\item{TFGenes}{genes regulated by the TFs as returned by simpleMotifTFMatch()
or findRegulatoryNetworks()}

\item{RNACoGAPSResult}{CoGAPSResult object from matched RNA-seq data, or, if
matrix = TRUE, a matrix containing patternMarker gene ranks}

\item{ATACPatternSet}{vector of patterns found by CoGAPS in the ATAC data to
match against patterns found in RNA}

\item{RNAPatternSet}{vector of patterns found by CoGAPS in RNA to match
against those found in ATAC}

\item{matrix}{TRUE if inputting matrix of PatternMarker genes, FALSE if
inputting CoGAPS result object. FALSE by default}
}
\value{
Result matrices from the fgsea function for each pattern comparison
}
\description{
Use results from CoGAPS run on matched RNA-seq data to verify TF activity
suggested by motif matching analysis of ATAC CoGAPS output. Uses the fgsea
package to find enrichment of PatternMarker genes among genes regulated by
identified candidate TFs
}
\examples{
\dontrun{
gseaList = RNAseqTFValidation(TFMatchResult$RegulatoryNetworks, RNACoGAPS, c(1,3), c(2,7), matrix = FALSE)
}
}
