% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schepCellTypes}
\alias{schepCellTypes}
\title{Cell types corresponding to subsetSchepData}
\format{
Factor of length 600 with 12 levels
}
\source{
\url{10.1038/nmeth.4401}
}
\usage{
schepCellTypes
}
\description{
Factor of cell types in the order of the subsetSchepData object from the
Schep et al, 2017, Nature Methods paper.
}
\keyword{datasets}
