% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schepCogapsResult}
\alias{schepCogapsResult}
\title{CogapsResult from the subsetSchepData object}
\format{
Large CogapsResult
}
\usage{
schepCogapsResult
}
\description{
Output from applying the CoGAPS algorithm to the subsetSchepData object.
}
\examples{
data(schepCogapsResult)
}
\keyword{datasets}
