% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schepGranges}
\alias{schepGranges}
\title{GRanges corresponding to subsetSchepData}
\format{
GRanges of length 5036
}
\source{
\url{10.1038/nmeth.4401}
}
\usage{
schepGranges
}
\description{
GRanges in the order of the peaks of the subsetSchepData object from the
Schep et al, 2017, Nature Methods paper.
}
\keyword{datasets}
