% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMotifTFMatch.R
\name{simpleMotifTFMatch}
\alias{simpleMotifTFMatch}
\title{Motif/TF Matching in a Single Function}
\usage{
simpleMotifTFMatch(
  cogapsResult,
  generanges,
  organism,
  genome,
  scoreThreshold = NULL,
  motifsPerRegion = 1
)
}
\arguments{
\item{cogapsResult}{result object from CoGAPS run}

\item{generanges}{GRanges object corresponding to peaks in ATACseq data
CoGAPS was run on}

\item{organism}{organism name (e.g. "Homo sapiens")}

\item{genome}{genome version to use (e.g. hg19, mm10)}

\item{scoreThreshold}{threshold for the most pattern defining peaks as per
the PatternMarker statistic from the CoGAPS package. By default is NULL, in
which case all Pattern defining peaks will be used for motif matching. Used
to reduce compute time, as results are quite robust across thresholds}

\item{motifsPerRegion}{number of motifs to attempt to find within each peak}
}
\value{
list containing list of matched motifs, list of transciption factors,
  regulatory gene networks known for those TFs, functional annotations,
  summary showing how many times each TF was matched to a peak, and the
  downloaded set of motifs for the user to save for reproducibility
}
\description{
If the user does not have a specific set of motifs, transcription factors, or
regulatory networks that they want to match against, simply uses the core
motifs from the JASPAR database to find motifs and TFs in the most Pattern
differentiating peaks, as well as regulatory networks from TTrust database
corresponding to the identified TFs. This is used to provide transcription
factors with functional annotation which may suggest plausible unknown
regulatory mechanisms operating in the cell types of interest within the
data.
}
\examples{
data("schepCogapsResult")
data(schepGranges)

motifResults = simpleMotifTFMatch(schepCogapsResult,
 generanges = schepGranges, organism = "Homo sapiens", genome = "hg19", scoreThreshold = 0.05)
}
