% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subsetSchepData}
\alias{subsetSchepData}
\title{Small subset of the scATAC-seq data from Schep et al, 2017, Nature Methods paper.}
\format{
A matrix with 5036 peaks and 600 cells in the order of the schepPeaks,
 schepCellTypes, and schepGranges data objects.
}
\source{
\url{10.1038/nmeth.4401}
}
\usage{
subsetSchepData
}
\description{
Subset from the Schep et al data, used for examples in this package.
}
\keyword{datasets}
