% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tfData}
\alias{tfData}
\title{List of human TFs and motifs from cisBP database}
\format{
Data frame with 95413 rows and 28 columns.
}
\source{
\url{http://cisbp.ccbr.utoronto.ca/}
}
\usage{
tfData
}
\description{
Information on human TFs and their correpsonding DNA motifs
}
\examples{
data(tfData)
}
\keyword{datasets}
