% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePeakMatchingv2.R
\name{genePatternMatch}
\alias{genePatternMatch}
\title{Match genes to pattern differentiating peaks}
\usage{
genePatternMatch(cogapsResult, generanges, genome, scoreThreshold = NULL)
}
\arguments{
\item{cogapsResult}{the CogapsResult object produced by a CoGAPS run}

\item{generanges}{GRanges object corresponding to the genomic regions
identified as peaks for the ATAC-seq data that CoGAPS was run on}

\item{genome}{A TxDb object for the genome of interest, it must be loaded
prior to calling this function}

\item{scoreThreshold}{threshold for the most pattern defining peaks as per
the PatternMarker statistic from the CoGAPS package. Default is NULL,
returning all PatternMarker peaks. Useful to reduce computational time, as
top results are reasonably robust to using more stringent thresholds}
}
\value{
double nested list containing lists of the genes in, nearest, and
  following the peaks matched each pattern
}
\description{
Function to take as input CoGAPS results for ATAC-seq data and find genes
within the most "pattern-defining" regions (as identified by cut thresholded
pattern Marker statistic from the CoGAPS package), as well as the nearest
gene and the nearest gene following the region. Note: a TxDb object for the
genome of interest must be loaded prior to running this function.
}
\examples{
data("schepCogapsResult")
data(schepGranges)

library(Homo.sapiens)

genes = genePatternMatch(cogapsResult = schepCogapsResult,
 generanges = schepGranges, genome = Homo.sapiens)
}
