% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_peak_matching.R
\name{genePatternMatch}
\alias{genePatternMatch}
\title{Match genes to pattern differentiating peaks}
\usage{
genePatternMatch(cogapsResult, generanges, genome, scoreThreshold = 0.03)
}
\arguments{
\item{cogapsResult}{the CogapsResult object produced by a CoGAPS run}

\item{generanges}{GRanges object corresponding to the genomic regions
identified as peaks for the ATAC-seq data that CoGAPS was run on}

\item{genome}{Homo.sapiens or Mus.musculus currently supported}

\item{scoreThreshold}{threshold for the most pattern defining peaks as per
the PatternMarker statistic from the CoGAPS package}
}
\value{
double nested list containing lists of the genes in, nearest, and
  following the peaks matched each pattern
}
\description{
Function to take as input CoGAPS results for ATAC-seq data and find genes
within the most "pattern-defining" regions (as identified by cut thresholded
pattern Marker statistic from the CoGAPS package), as well as the nearest
gene and the nearest gene following the region
}
\examples{
data(schepCogapsResult)
data(schepGranges)

genes = genePatternMatch(cogapsResult = schepCogapsResult, generanges = schepGranges, genome = Homo.sapiens, scoreThreshold = 0.03)
}
