% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifPatternMatch.R
\name{motifPatternMatch}
\alias{motifPatternMatch}
\alias{getTFs}
\alias{findRegulatoryNetworks}
\alias{getTFDescriptions}
\title{Find Motifs and TFs from PatternMarker Peaks}
\usage{
motifPatternMatch(
  cogapsResult,
  generanges,
  motiflist,
  genome,
  scoreThreshold = NULL,
  motifsPerRegion = 1
)

getTFs(motifList, tfData)

findRegulatoryNetworks(TFs, networks)

getTFDescriptions(TFs)
}
\arguments{
\item{cogapsResult}{the result object from a CoGAPS run}

\item{generanges}{GRanges objects corresponding to the genomic regions which
form the rows of the ATAC-seq data that CoGAPS was run on}

\item{motiflist}{a PWMlist of motifs to search the regions for}

\item{genome}{the ucsc genome version to use e.g. "hg19", "mm10"}

\item{scoreThreshold}{threshold for the most pattern defining peaks as per
the PatternMarker statistic from the CoGAPS package. By default is NULL, in
which case all Pattern defining peaks will be used for motif matching. Used
to reduce compute time, as results are quite robust across thresholds}

\item{motifsPerRegion}{number of top motifs to return from each peak}

\item{motifList}{list produced by the motifPatternMatch function}

\item{tfData}{dataframe of motifs and TFs from cisBP database}

\item{TFs}{object of TF info returned from the getTFs function}

\item{networks}{a list of regulatory networks of genes corresponding to TFs;
we include humanRegNets and mouseRegNets, downloaded from the TTrust
database (Han et al Nucleic Acid Res. 2018)}
}
\value{
motifPatternMatch: nested list of the top motif for each region for x
  number of regions for each pattern

getTFs: list containing list of dataframes of tfData subset to matched TFs
  and list of how many times each TF was matched to a motif/peak

findRegulatoryNetworks: list of TFs for which we have annotations and
  the corresponding gene networks for each pattern

getTFDescriptions: list of functional annotations for all TFs in each
  pattern
}
\description{
Function that takes CoGAPS result and list of DNA motifs as input and returns
motifs which match to the most pattern-defining peaks for each pattern.
}
\section{Functions}{
\itemize{
\item \code{getTFs}: Match motifs to TFs based on the list of motifs
returned by motifPatternMatch

\item \code{findRegulatoryNetworks}: function to match TFs identified by getTFs
function to a list of regulatory networks of genes known for those TFs

\item \code{getTFDescriptions}: function to match functional annotation to a
list of TFs from the getTFs function
}}

\examples{
data(exampleMotifList)
data(schepGranges)
data(schepCogapsResult)

motifsByPattern = motifPatternMatch(schepCogapsResult, schepGranges,
 exampleMotifList, "hg19")
data(exampleMotifList)
data(schepGranges)
data(schepCogapsResult)
data(tfData)

motifsByPattern = motifPatternMatch(schepCogapsResult, schepGranges, exampleMotifList, "hg19")
motifTFs = getTFs(motifsByPattern, tfData)
data(exampleMotifList)
data(schepGranges)
data(schepCogapsResult)
data(tfData)

motifsByPattern = motifPatternMatch(schepCogapsResult, schepGranges, exampleMotifList, "hg19")
motifTFs = getTFs(motifsByPattern, tfData)

regNets = findRegulatoryNetworks(motifTFs, ATACCoGAPS:::humanRegNets)
data(exampleMotifList)
data(schepGranges)
data(schepCogapsResult)
data(tfData)

motifsByPattern = motifPatternMatch(schepCogapsResult, schepGranges, exampleMotifList, "hg19")
motifTFs = getTFs(motifsByPattern, tfData)

tfDesc = getTFDescriptions(motifTFs)
}
